requirements.txt
langchain>=1.2.0
langgraph>=1.1.0
langchain-openai>=1.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
